# wise2022-webshop/item-sb/OpenAPI.yaml
openapi: 3.0.1
info:
  title: HS Esslingen Webshop
  description: 'This is a sample webshop to demonstrate a fullstack web application'
  contact:
    email: joerg.nitzsche@hs-esslingen.de
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
externalDocs:
  description: Find out more about HS Esslingen
  url: http://hs-esslingen.de
servers:
- url: http://localhost:8200/api/item-management/v1
tags:
- name: item
  description: Everything about items
  externalDocs:
    description: Find out more 
    url: http://hs-esslingen.de
paths:
  /item/{itemId}:
    get:
      tags:
      - item
      summary: Find item by ID
      description: Returns a single item
      operationId: getItemById
      parameters:
      - name: itemId
        in: path
        description: ID of item to return
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Item'
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        404:
          description: Item not found
          content: {}
    delete:
      tags:
      - item
      summary: Deletes an item
      operationId: deleteItem
      parameters:
      - name: itemId
        in: path
        description: Item id to delete
        required: true
        schema:
          type: integer
          format: int64
      responses:
        204:
          description: no data
          content: {}
        404:
          description: Item not found
          content: {}
components:
  schemas:
    Item:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
      xml:
        name: Item